requirements.txt
tweepy 
pymongo 
requests 
pyyaml